tests: round out data test suite and fix bugs
