+++ tests/conftest.py
@@ -1,0 +1,6 @@
+def pytest_addoption(parser):
+    parser.addoption(
+        "--diff-lines",
+        action='store',
+        default=200,
+    )
